cocall(2): Return receive size instead of 0 on success
This makes it possible for callers to verify the other side
didn't send an undersized buffer, which could then cause leaking
data from a previous request.  This means we no longer need
to clear input buffers.

XXX With cocall_slow_test, there are broken cocall_proxy_abort
    and cocall_callee_dead, and flaky cocall_proxy.  Non-slow
    seems a bit more broken.

+++ sys/compat/freebsd64/freebsd64_systrace_args.c
@@ -11203,12 +11203,12 @@ systrace_return_setargdesc(int sysnum, int ndx, char *desc, size_t descsz)
 	/* freebsd64_cocall_slow */
 	case 548:
 		if (ndx == 0 || ndx == 1)
-			p = "int";
+			p = "ssize_t";
 		break;
 	/* freebsd64_coaccept_slow */
 	case 549:
 		if (ndx == 0 || ndx == 1)
-			p = "int";
+			p = "ssize_t";
 		break;
 	/* fdatasync */
 	case 550:
